pyproject.toml
[tool.poetry]
name = "turbo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11.0"
redis = "^4.4.2"
gurobipy = "^10.0.0"
loguru = "^0.6.0"
numpy = "^1.24.1"
opacus = "^1.3.0"
omegaconf = "^2.3.0"
pandas = "^1.5.2"
plotly = "^5.12.0"
psycopg2 = "^2.9.5"
PyYAML = "^6.0"
typer = "^0.7.0"
termcolor = "^2.2.0"
sortedcollections = "^2.1.0"
black = "^22.12.0"
simpy = "^4.0.1"
jupyter = "^1.0.0"
nb-clean = "^2.4.0"
modin = "^0.18.1"
setuptools = "^67.1.0"
geopy = "^2.3.0"
redisai = "^1.3.0"
threaded = "^4.1.0"
kaleido = "0.2.1"
scalene = "^1.5.20"
grpcio = "<=1.48.2"
mlflow = "^2.3.1"
ray = "^2.4.0"
colorama = "^0.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
